Add open settings

+++ DocumentManagement.Launcher/NotifyIconViewModel.cs
@@ -23,8 +23,12 @@ namespace DocumentManagement.Launcher
             VisibleConsoleCommand = new HCommand(VisibleConsole);
             ExitApplicationCommand = new HCommand(ExitApplication);
             OpenSwaggerCommand = new HCommand(OpenSwagger);
+            StartDmConsoleCommand = new HCommand(StartDocumentMenagermentProcess);
+            OpenSettingsCommand = new HCommand(OpenSettings);
             StartDocumentMenagermentProcess();
         }
+
+        
         #endregion
 
         #region binding
@@ -44,6 +48,10 @@ namespace DocumentManagement.Launcher
 
         public HCommand OpenSwaggerCommand { get; }
 
+        public HCommand StartDmConsoleCommand { get; }
+
+        public HCommand OpenSettingsCommand { get; }
+
         public bool IsVisibleConsole
         {
             get => LauncherSettings.Instance.VisibleConsole;
@@ -57,13 +65,22 @@ namespace DocumentManagement.Launcher
 
         public void Dispose()
         {
-            dmProcess?.Kill();
-            dmProcess?.Dispose();
+            if (dmProcess != null)
+            {
+                dmProcess?.Kill();
+                dmProcess?.WaitForExit();
+                dmProcess?.Dispose();
+            }
         }
 
         #region private method
-        [DllImport("user32.dll")]
-        private static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
+
+        private void OpenSettings()
+        {
+            var notepad = Process.Start("notepad", LauncherSettings.PATH);
+            notepad.EnableRaisingEvents = true;
+            notepad.Exited += (s, o) => LauncherSettings.Reload();
+        }
 
         private void OpenSwagger() => OpenUrl(LauncherSettings.Instance.SwaggerPath);
 
@@ -86,25 +103,26 @@ namespace DocumentManagement.Launcher
 
             dmProcess = new Process();
             dmProcess.StartInfo.FileName = path;
-            dmProcess.StartInfo.CreateNoWindow = !IsVisibleConsole;
-            IsRunDm = true;
-            dmProcess.Start();
+            dmProcess.StartInfo.CreateNoWindow = false;
+            // TODO: dmProcess.StartInfo.Arguments =
             dmProcess.EnableRaisingEvents = true;
             dmProcess.Exited += DmProcessDisposed;
             dmProcess.Disposed += DmProcessDisposed;
+            IsRunDm = true;
+            dmProcess.Start();
+            if (!IsVisibleConsole)
+                WindowOperation.Hide(dmProcess.MainWindowHandle);
         }
 
         private void VisibleConsole()
         {
-            if (dmProcess == null)
-            {
+            if (dmProcess == null)            
                 StartDocumentMenagermentProcess();
-            }
-            else
-            {
-                ShowWindow(dmProcess.MainWindowHandle, IsVisibleConsole ? 0 : 5);
-                IsVisibleConsole = !IsVisibleConsole;
-            }
+            IsVisibleConsole = !IsVisibleConsole;
+            if (IsVisibleConsole)
+                WindowOperation.Show(dmProcess.MainWindowHandle);
+            else 
+                WindowOperation.Hide(dmProcess.MainWindowHandle);
         }
 
         private void ExitApplication()
@@ -141,5 +159,28 @@ namespace DocumentManagement.Launcher
             }
         }
         #endregion
+    }
+
+    public static class WindowOperation
+    {
+        public static void Hide(IntPtr win)
+        {
+            ShowWindow(win, 0);
+        }
+
+        public static void Show(IntPtr win)
+        {
+            ShowWindow(win, 1);
+        }
+
+        // Link: https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-showwindow
+        [DllImport("user32.dll")]
+        private static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
+
+        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+        private static extern IntPtr FindWindow(string lpClassName, string lpWindowName);
+
+        [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+        private static extern IntPtr FindWindowEx(IntPtr parentHandle, IntPtr childAfter, string lclassName, string windowTitle);
     }
 }
